+++ WebApplication/src/store/modules/labels.js
@@ -6,7 +6,8 @@ import store from "../../store"
 const state = () => ({
     allLabels: [],    // all labels in Orthanc
     loaded: false,
-    canShowStudiesWithoutLabels: false
+    canShowStudiesWithoutLabels: false,
+    hasAccessToAllLabels: false
 })
 
 ///////////////////////////// GETTERS
@@ -24,6 +25,9 @@ const mutations = {
     },
     setCanShowStudiesWithoutLabels(state) {
         state.canShowStudiesWithoutLabels = true;
+    },
+    setHasAccessToAllLabels(state) {
+        state.hasAccessToAllLabels = true;
     }
 }
 
@@ -35,15 +39,21 @@ const actions = {
         commit('setLabels', { allLabels: allLabels });
 
         let showStudiesWithoutLabels = false;
+        let hasAccessToAllLabels = true;
+
         if (store.state.configuration.system.ApiVersion >= 30 && allLabels.length > 0) {
             if (store.state.configuration.userProfile) { // only the users who have access to all labels can see this option !
-                showStudiesWithoutLabels = store.state.configuration.userProfile["authorized-labels"].length == 1 && store.state.configuration.userProfile["authorized-labels"][0] == "*";
+                hasAccessToAllLabels = store.state.configuration.userProfile["authorized-labels"].length == 1 && store.state.configuration.userProfile["authorized-labels"][0] == "*";
+                showStudiesWithoutLabels = hasAccessToAllLabels;
             } else {
                 showStudiesWithoutLabels = true;
             }
             if (showStudiesWithoutLabels) {
                 commit('setCanShowStudiesWithoutLabels');
             }
+        }
+        if (hasAccessToAllLabels) {
+            commit('setHasAccessToAllLabels');
         }
 
         commit('setLoaded');